updated failures and refactoring
